requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.7.4
chardet==4.0.0
idna==2.10
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
requests==2.25.1
six==1.16.0
tornado==6.4.1
tzlocal==5.2
urllib3==1.26.19
